// src/utils/types/types.ts
export interface ICharacter {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export interface IEpisode {
  id: number;
  name: string;
  air_date: string;
  episode: string;
  characters: string[];
  url: string;
  created: string;
}

export interface ILocation {
  id: number;
  name: string;
  type: string;
  dimension: string;
  residents: string[];
  url: string;
  created: string;
}

export interface ILocations {
  info: {
    count: number;
    pages: number;
    next: string;
    prev: string | null;
  };
  results: ILocation[];
}

export interface ICharacters {
  info: {
    count: number;
    pages: number;
    next: string;
    prev: string | null;
  };
  results: ICharacter[];
}
export interface IEpisodes {
  info: {
    count: number;
    pages: number;
    next: string;
    prev: string | null;
  };
  results: IEpisode[];
}